requirements.txt
flask
pillow
opencv-python
pytesseract
pyyaml
tensorflow

